// web/src/ar/pages/version-details/DartVersion/types.ts
/*
 * Copyright 2024 Harness, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import type { ArtifactDetail } from '@harnessio/react-har-service-client'
import type { VersionDetailsQueryParams } from '../types'

export enum DartArtifactDetailsTabEnum {
  ReadMe = 'readme',
  Files = 'files',
  Dependencies = 'dependencies'
}
export interface DartVersionDetailsQueryParams extends VersionDetailsQueryParams {
  detailsTab: DartArtifactDetailsTabEnum
}

export type LocalDartArtifactDetailConfig = {
  metadata?: {
    description?: string
    documentation_url?: string
    project_url?: string
    repository_url?: string
    readme?: string
    pubspec?: {
      dependencies?: Record<string, string>
      dev_dependencies?: Record<string, string>
      home_page?: string
      name?: string
      version?: string
    }
  }
}

export type DartArtifactDetails = ArtifactDetail & LocalDartArtifactDetailConfig